Define InflateColumn::Bytea
When retrieving bytea fields from the database, we should decode them
from utf-8 to characters. The way to the database is more complicated:
Depending on the way this report data is posted to the service we need
to add an extra encode() for those bytea fields. That is not done here.